Add a user namespace
To avoid 'operation not permitted' errors with runC v0.0.3, I had to
drop the tmpfs, devpts, and sysfs mounts entirely, as well as dropping
the 'ro' limit from the cgroup mount.

Now that we're using user namespaces, we can drop the chmod rules from
the Makefile.

To fill in the local user's user and group IDs, I'm using gettext's
envsubst [1].  It's not in POSIX [2], so it may be slightly less
portable.  However, envsubst and gettext are in Gentoo's x86 and amd64
stage1s, so it can't be *that* uncommon ;).

[1]: http://www.gnu.org/software/gettext/manual/html_node/envsubst-Invocation.html#envsubst-Invocation
[2]: http://pubs.opengroup.org/onlinepubs/9699919799/idx/utilities.html